requirements.txt
Django
gunicorn
psycopg2
djangorestframework
jsonfield
django-parsley
celery
redis
